+++ src/components/common/bottom-sheet-base/BottomSheetBase.ts
@@ -1,4 +1,4 @@
-import { EVENT_TYPES } from "../../../constants/constants.js";
+import { EVENT_TYPES } from "../../../constants";
 import "./bottomSheetBase.css";
 
 interface BottomSheetBaseProps {
